csi: allow building without cgo dependencies for linting
